Get question names from an XLSForm

Returns the values of the name column from the survey sheet, excluding any rows where name is NA (such as begin_group / end_group rows that carry no name).

Usage

xlsform_questions(x, ...)

# Default S3 method
xlsform_questions(x, ...)

# S3 method for class 'xlsform'
xlsform_questions(x, ...)

Arguments

x

An xlsform object.

...

Ignored; present for S3 method compatibility.

Value

A character vector of non-NA question names from the survey sheet.

Details

The returned vector is used internally by validate_question_names() to compare question inventories across two forms.

See also

Examples

form <- read_xlsform(system.file("extdata/form.xlsx", package = "idem"))

# All question names in the form
xlsform_questions(form)
#>   [1] "audit"                                                     
#>   [2] "start"                                                     
#>   [3] "end"                                                       
#>   [4] "today"                                                     
#>   [5] "deviceid"                                                  
#>   [6] "instance_name"                                             
#>   [7] "introduction"                                              
#>   [8] "survey_modality"                                           
#>   [9] "enum_id"                                                   
#>  [10] "enum_gender"                                               
#>  [11] "other_enum_gender"                                         
#>  [12] "admin1"                                                    
#>  [13] "admin2"                                                    
#>  [14] "admin3"                                                    
#>  [15] "admin4"                                                    
#>  [16] "cluster_id"                                                
#>  [17] "consent"                                                   
#>  [18] "consent_no_note"                                           
#>  [19] "intro_hh"                                                  
#>  [20] "introduction"                                              
#>  [21] "consented"                                                 
#>  [22] "demographics"                                              
#>  [23] "resp_gender"                                               
#>  [24] "resp_age"                                                  
#>  [25] "resp_hoh_yn"                                               
#>  [26] "hoh_gender"                                                
#>  [27] "hoh_age"                                                   
#>  [28] "hoh_civil_status"                                          
#>  [29] "other_hoh_civil_status"                                    
#>  [30] "setting"                                                   
#>  [31] "hh_size"                                                   
#>  [32] "repeat_intro_hh"                                           
#>  [33] "roster"                                                    
#>  [34] "parent_instance_name"                                      
#>  [35] "person_id"                                                 
#>  [36] "ind_pos"                                                   
#>  [37] "note_resp"                                                 
#>  [38] "ind_name"                                                  
#>  [39] "ind_gender"                                                
#>  [40] "ind_age"                                                   
#>  [41] "check_gender"                                              
#>  [42] "check_age"                                                 
#>  [43] "note_resp_gender_mismatch"                                 
#>  [44] "note_resp_age_mismatch"                                    
#>  [45] "ind_age_under1"                                            
#>  [46] "ind_under5_date_know"                                      
#>  [47] "ind_under5_date"                                           
#>  [48] "ind_under5_event"                                          
#>  [49] "ind_dob_final"                                             
#>  [50] "ind_under5_age_months"                                     
#>  [51] "ind_under5_age_years"                                      
#>  [52] "note_child_age_under5"                                     
#>  [53] "check_age_under5"                                          
#>  [54] "note_child_age_under5_mismatch"                            
#>  [55] "ind_age_below18"                                           
#>  [56] "ind_age_2_11"                                              
#>  [57] "ind_f"                                                     
#>  [58] "ind_age_5_17"                                              
#>  [59] "ind_m_age_5_17"                                            
#>  [60] "ind_f_age_5_17"                                            
#>  [61] "ind_age_0_1"                                               
#>  [62] "ind_age_0_4"                                               
#>  [63] "ind_age_0_5"                                               
#>  [64] "ind_f_age_above18"                                         
#>  [65] "ind_m_age_above18"                                         
#>  [66] "ind_woman_repr_age"                                        
#>  [67] "ind_age_schooling"                                         
#>  [68] "ind_potentially_hoh"                                       
#>  [69] "wgss_indiv"                                                
#>  [70] "wgss_ind_pos"                                              
#>  [71] "wgss_person_id"                                            
#>  [72] "wgss_ind_age"                                              
#>  [73] "wgss_ind_gender"                                           
#>  [74] "wgss_ind_name"                                             
#>  [75] "direct_proxy"                                              
#>  [76] "caregiver_available"                                       
#>  [77] "proxy_resp_gender"                                         
#>  [78] "proxy_resp_age"                                            
#>  [79] "wgq_vision"                                                
#>  [80] "wgq_hearing"                                               
#>  [81] "wgq_mobility"                                              
#>  [82] "wgq_cognition"                                             
#>  [83] "wgq_self_care"                                             
#>  [84] "wgq_communication"                                         
#>  [85] "gi_dis_challenges"                                         
#>  [86] "wgss_indiv"                                                
#>  [87] "roster"                                                    
#>  [88] "ind_age_2_11_n"                                            
#>  [89] "ind_f_n"                                                   
#>  [90] "ind_age_5_17_n"                                            
#>  [91] "ind_m_age_5_17_n"                                          
#>  [92] "ind_f_age_5_17_n"                                          
#>  [93] "ind_age_0_1_n"                                             
#>  [94] "ind_age_0_4_n"                                             
#>  [95] "ind_age_0_5_n"                                             
#>  [96] "ind_f_age_above18_n"                                       
#>  [97] "ind_m_age_above18_n"                                       
#>  [98] "ind_age_above18_n"                                         
#>  [99] "ind_age_below18_n"                                         
#> [100] "ind_age_schooling_n"                                       
#> [101] "ind_women_repr_age_n"                                      
#> [102] "ind_hoh_mismatch"                                          
#> [103] "demographics"                                              
#> [104] "displacement"                                              
#> [105] "note_dis_hh"                                               
#> [106] "dis_forced"                                                
#> [107] "dis_area_origin"                                           
#> [108] "dis_reasons"                                               
#> [109] "other_dis_reasons"                                         
#> [110] "dis_main_reasons"                                          
#> [111] "other_dis_main_reasons"                                    
#> [112] "dis_duration"                                              
#> [113] "dis_secondary_n"                                           
#> [114] "displacement"                                              
#> [115] "aap"                                                       
#> [116] "cpa_priority_challenge_note"                               
#> [117] "cpa_preferred_modality_1"                                  
#> [118] "other_cpa_preferred_modality_1"                            
#> [119] "cpa_priority_support_ngo_1"                                
#> [120] "other_cpa_priority_support_ngo_1"                          
#> [121] "cpa_preferred_modality_2"                                  
#> [122] "other_cpa_preferred_modality_2"                            
#> [123] "cpa_priority_support_ngo_2"                                
#> [124] "other_cpa_priority_support_ngo_2"                          
#> [125] "cpa_priority_support_cash"                                 
#> [126] "aap_received_assistance_12m"                               
#> [127] "aap_received_assistance_date"                              
#> [128] "aap_received_assistance_type"                              
#> [129] "other_aap_received_assistance_type"                        
#> [130] "aap_relevance_assistance"                                  
#> [131] "aap_relevance_assistance_reason"                           
#> [132] "other_aap_relevance_assistance_reason"                     
#> [133] "aap_satisfaction_assistance"                               
#> [134] "aap_satisfaction_challenges"                               
#> [135] "other_aap_satisfaction_challenges"                         
#> [136] "aap_assistance_improves_living_conditions"                 
#> [137] "other_aap_assistance_improves_living_conditions"           
#> [138] "aap_assistance_improves_living_conditions_challenges"      
#> [139] "other_aap_assistance_improves_living_conditions_challenges"
#> [140] "aap_assistance_coverage"                                   
#> [141] "aap_satisfaction_workers_behaviour"                        
#> [142] "aap_satisfaction_workers_behaviour_reason"                 
#> [143] "other_aap_satisfaction_workers_behaviour_reason"           
#> [144] "aap_consultation_assistance_yn"                            
#> [145] "aap_consultation_assistance_opinions"                      
#> [146] "aap_preferred_means_feedback"                              
#> [147] "other_aap_preferred_means_feedback"                        
#> [148] "aap_cfm_channels_awareness"                                
#> [149] "other_aap_cfm_channels_awareness"                          
#> [150] "aap_cfm_use"                                               
#> [151] "aap_information_needs"                                     
#> [152] "other_aap_information_needs"                               
#> [153] "aap_preferred_source_info"                                 
#> [154] "other_aap_preferred_source_info"                           
#> [155] "aap_received_type_info_aid"                                
#> [156] "other_aap_received_type_info_aid"                          
#> [157] "aap_received_source_info_aid"                              
#> [158] "other_aap_received_source_info_aid"                        
#> [159] "aap_received_channel_info_aid"                             
#> [160] "other_aap_received_channel_info_aid"                       
#> [161] "aap_trusted_channel_info"                                  
#> [162] "other_aap_trusted_channel_info"                            
#> [163] "hesper"                                                    
#> [164] "hesper_drinking_water"                                     
#> [165] "hesper_food"                                               
#> [166] "hesper_shelter"                                            
#> [167] "hesper_toilet"                                             
#> [168] "hesper_clean"                                              
#> [169] "hesper_clean_women"                                        
#> [170] "hesper_clothes_etc"                                        
#> [171] "hesper_income_livelihood"                                  
#> [172] "hesper_health"                                             
#> [173] "hesper_health_care_men"                                    
#> [174] "hesper_health_care_women"                                  
#> [175] "hesper_distress"                                           
#> [176] "hesper_safety"                                             
#> [177] "hesper_education"                                          
#> [178] "hesper_care"                                               
#> [179] "hesper_support"                                            
#> [180] "hesper_separation"                                         
#> [181] "hesper_displaced"                                          
#> [182] "hesper_information"                                        
#> [183] "hesper_aid"                                                
#> [184] "hesper_respect"                                            
#> [185] "hesper_movement"                                           
#> [186] "hesper_time"                                               
#> [187] "hesper_community_intro"                                    
#> [188] "hesper_law"                                                
#> [189] "hesper_gbv"                                                
#> [190] "hesper_drug"                                               
#> [191] "hesper_mental_health"                                      
#> [192] "hesper_care_community"                                     
#> [193] "hesper_priority_intro"                                     
#> [194] "hesper_priority_first"                                     
#> [195] "hesper_priority_second"                                    
#> [196] "hesper_priority_third"                                     
#> [197] "hesper"                                                    
#> [198] "aap"                                                       
#> [199] "edu"                                                       
#> [200] "edu_ind_note"                                              
#> [201] "edu_ind"                                                   
#> [202] "edu_ind_pos"                                               
#> [203] "edu_person_id"                                             
#> [204] "edu_ind_age"                                               
#> [205] "edu_ind_gender"                                            
#> [206] "edu_ind_age_schooling"                                     
#> [207] "edu_ind_name"                                              
#> [208] "edu_ind_gender_en"                                         
#> [209] "edu_ind_gender_fr"                                         
#> [210] "edu_access"                                                
#> [211] "edu_level_grade"                                           
#> [212] "edu_disrupted_hazards"                                     
#> [213] "edu_disrupted_teacher"                                     
#> [214] "edu_disrupted_displaced"                                   
#> [215] "edu_disrupted_attack"                                      
#> [216] "edu_barrier"                                               
#> [217] "other_edu_barrier"                                         
#> [218] "edu_financial_barrier"                                     
#> [219] "other_edu_financial_barrier"                               
#> [220] "edu_other_yn"                                              
#> [221] "edu_other_type"                                            
#> [222] "edu_ind"                                                   
#> [223] "edu_community_modality"                                    
#> [224] "edu"                                                       
#> [225] "shelter"                                                   
#> [226] "snfi_note"                                                 
#> [227] "snfi_shelter_type"                                         
#> [228] "snfi_shelter_type_collective"                              
#> [229] "other_snfi_shelter_type_collective"                        
#> [230] "snfi_shelter_type_individual"                              
#> [231] "other_snfi_shelter_type_individual"                        
#> [232] "snfi_shelter_damage"                                       
#> [233] "other_snfi_shelter_damage"                                 
#> [234] "snfi_shelter_damage_when"                                  
#> [235] "snfi_shelter_damage_cause"                                 
#> [236] "other_snfi_shelter_damage_cause"                           
#> [237] "snfi_shelter_damage_barriers_repairs"                      
#> [238] "snfi_shelter_issue"                                        
#> [239] "other_snfi_shelter_issue"                                  
#> [240] "snfi_fds_cooking"                                          
#> [241] "snfi_fds_cooking_issue"                                    
#> [242] "other_snfi_fds_cooking_issue"                              
#> [243] "snfi_fds_sleeping"                                         
#> [244] "snfi_fds_sleeping_issue"                                   
#> [245] "other_snfi_fds_sleeping_issue"                             
#> [246] "snfi_fds_storing"                                          
#> [247] "snfi_fds_storing_issue"                                    
#> [248] "other_snfi_fds_storing_issue"                              
#> [249] "snfi_sufficient_clothes"                                   
#> [250] "snfi_essential_items_missing"                              
#> [251] "other_snfi_essential_items_missing"                        
#> [252] "shelter"                                                   
#> [253] "energy"                                                    
#> [254] "energy_lighting_source"                                    
#> [255] "other_energy_lighting_source"                              
#> [256] "energy"                                                    
#> [257] "hlp"                                                       
#> [258] "hlp_occupancy"                                             
#> [259] "other_hlp_occupancy"                                       
#> [260] "hlp_risk_eviction"                                         
#> [261] "hlp_threat_eviction"                                       
#> [262] "hlp"                                                       
#> [263] "wash"                                                      
#> [264] "wash_note"                                                 
#> [265] "wash_drinking_water_source"                                
#> [266] "other_wash_drinking_water_source"                          
#> [267] "wash_drinking_water_time_yn"                               
#> [268] "wash_drinking_water_time_int"                              
#> [269] "wash_drinking_water_time_sl"                               
#> [270] "wash_hwise_drink"                                          
#> [271] "wash_hwise_hands"                                          
#> [272] "wash_hwise_worry"                                          
#> [273] "wash_hwise_plans"                                          
#> [274] "wash_hwise_12_none"                                        
#> [275] "wash_hwise_12_sleep"                                       
#> [276] "wash_hwise_12_food"                                        
#> [277] "wash_hwise_12_clothes"                                     
#> [278] "wash_hwise_12_angry"                                       
#> [279] "wash_hwise_12_body"                                        
#> [280] "wash_hwise_12_shame"                                       
#> [281] "wash_hwise_12_interrupt"                                   
#> [282] "wash_drinking_water_quantity"                              
#> [283] "wash_drinking_water_acceptable"                            
#> [284] "wash_water_access_issue"                                   
#> [285] "other_wash_water_access_issue"                             
#> [286] "wash_drinking_water_other_sources"                         
#> [287] "other_wash_drinking_water_other_sources"                   
#> [288] "wash_drinking_water_person_fetch"                          
#> [289] "wash_drinking_water_person_trips"                          
#> [290] "wash_drinking_water_piped_supply"                          
#> [291] "wash_water_availability_yn"                                
#> [292] "wash_water_availability_issue"                             
#> [293] "other_wash_water_availability_issue"                       
#> [294] "wash_drinking_water_supplied_hours"                        
#> [295] "wash_drinking_water_days_unavailable"                      
#> [296] "wash_drinking_water_store_insufficient_yn"                 
#> [297] "wash_drinking_water_tank_yn"                               
#> [298] "wash_drinking_water_tank_litres"                           
#> [299] "wash_drinking_water_tank_times_filled"                     
#> [300] "wash_drinking_water_store_small_containers"                
#> [301] "wash_drinking_water_safer_yn"                              
#> [302] "wash_drinking_water_safer_details"                         
#> [303] "other_wash_drinking_water_safer_details"                   
#> [304] "wash_sanitation_facility"                                  
#> [305] "other_wash_sanitation_facility"                            
#> [306] "wash_sanitation_facility_sharing_yn"                       
#> [307] "wash_sanitation_facility_sharing_n"                        
#> [308] "wash_sanitation_sharing_public"                            
#> [309] "wash_sanitation_toilet_design"                             
#> [310] "wash_sanitation_access_features"                           
#> [311] "wash_sanitation_facility_location"                         
#> [312] "wash_sanitation_use_of_facility"                           
#> [313] "wash_sanitation_use_toilet"                                
#> [314] "wash_sanitation_use_toilet_reason"                         
#> [315] "other_wash_sanitation_use_toilet_reason"                   
#> [316] "wash_sanitation_toilet_risks"                              
#> [317] "other_wash_sanitation_toilet_risks"                        
#> [318] "wash_sanitation_access_issue"                              
#> [319] "other_wash_sanitation_access_issue"                        
#> [320] "wash_sanitation_emptied_yn"                                
#> [321] "wash_sanitation_emptied_where"                             
#> [322] "other_wash_sanitation_emptied_where"                       
#> [323] "wash_sanitation_outlet_pipe"                               
#> [324] "wash_sanitation_outlet_pipe_where"                         
#> [325] "other_wash_sanitation_outlet_pipe_where"                   
#> [326] "wash_sanitation_facility_leak"                             
#> [327] "wash_sanitation_septic_discharge"                          
#> [328] "other_wash_sanitation_septic_discharge"                    
#> [329] "wash_sanitation_septic_age"                                
#> [330] "wash_sanitation_latrine_emptied"                           
#> [331] "wash_sanitation_emptied_who"                               
#> [332] "other_wash_sanitation_emptied_who"                         
#> [333] "wash_sanitation_excreta_released"                          
#> [334] "other_wash_sanitation_excreta_released"                    
#> [335] "wash_sanitation_adaptation"                                
#> [336] "other_wash_sanitation_adaptation"                          
#> [337] "wash_sanitation_environment"                               
#> [338] "wash_sanitation_dispose_garbage"                           
#> [339] "wash_sanitation_dispose_water"                             
#> [340] "wash_handwashing_facility"                                 
#> [341] "other_wash_handwashing_facility"                           
#> [342] "wash_handwashing_facility_observed_water_yn"               
#> [343] "wash_handwashing_facility_reported"                        
#> [344] "other_wash_handwashing_facility_reported"                  
#> [345] "wash_handwashing_facility_water_reported_yn"               
#> [346] "wash_soap_observed_yn"                                     
#> [347] "wash_soap_observed_type"                                   
#> [348] "other_wash_soap_observed_type"                             
#> [349] "wash_soap_reported_yn"                                     
#> [350] "wash_soap_reported_type"                                   
#> [351] "other_wash_soap_reported_type"                             
#> [352] "wash_hygiene_menstrual_privacy_yn"                         
#> [353] "wash_hygiene_menstrual_materials"                          
#> [354] "other_wash_hygiene_menstrual_materials"                    
#> [355] "wash_hygiene_menstrual_materials_reusable_yn"              
#> [356] "wash_hygiene_menstrual_missed_activities"                  
#> [357] "wash_hygiene_menstrual_missed_social_yn"                   
#> [358] "wash_hygiene_menstrual"                                    
#> [359] "other_wash_hygiene_menstrual"                              
#> [360] "wash_hygiene_menstrual_enough_materials_yn"                
#> [361] "wash_hygiene_menstrual_issue"                              
#> [362] "other_wash_hygiene_menstrual_issue"                        
#> [363] "wash_hygiene_menstrual_preferred"                          
#> [364] "other_wash_hygiene_menstrual_preferred"                    
#> [365] "wash_hygiene_menstrual_place"                              
#> [366] "wash_hygiene_adaptation"                                   
#> [367] "other_wash_hygiene_adaptation"                             
#> [368] "wash_bathing_facility_yn"                                  
#> [369] "wash_hygiene_nfi"                                          
#> [370] "other_wash_hygiene_nfi"                                    
#> [371] "wash_bathing_facility_issue"                               
#> [372] "other_wash_bathing_facility_issue"                         
#> [373] "wash_sanqol5_disease"                                      
#> [374] "wash_sanqol5_disg"                                         
#> [375] "wash_sanqol5_seen"                                         
#> [376] "wash_sanqol5_shame"                                        
#> [377] "wash_sanqol5_unsafe"                                       
#> [378] "wash_sanqol5_od_disease"                                   
#> [379] "wash_sanqol5_od_disg"                                      
#> [380] "wash_sanqol5_od_seen"                                      
#> [381] "wash_sanqol5_od_shame"                                     
#> [382] "wash_sanqol5_od_unsafe"                                    
#> [383] "wash_water_adaptation"                                     
#> [384] "other_wash_water_adaptation"                               
#> [385] "wash_productive_water_quantity"                            
#> [386] "wash_chl_info"                                             
#> [387] "wash_chl_info_src"                                         
#> [388] "other_wash_chl_info_src"                                   
#> [389] "wash_chl_actions_sick"                                     
#> [390] "other_wash_chl_actions_sick"                               
#> [391] "wash_chl_origin"                                           
#> [392] "other_wash_chl_origin"                                     
#> [393] "wash_chl_transmission"                                     
#> [394] "other_wash_chl_transmission"                               
#> [395] "wash_chl_health_actions"                                   
#> [396] "other_wash_chl_health_actions"                             
#> [397] "wash_chl_health_practices"                                 
#> [398] "other_wash_chl_health_practices"                           
#> [399] "wash_chl_ors"                                              
#> [400] "wash_chl_ors_preparation"                                  
#> [401] "other_wash_chl_ors_preparation"                            
#> [402] "wash"                                                      
#> [403] "fsl"                                                       
#> [404] "fsl_note"                                                  
#> [405] "fsl_respondent"                                            
#> [406] "fsl_fcs_cereal"                                            
#> [407] "fsl_fcs_legumes"                                           
#> [408] "fsl_fcs_dairy"                                             
#> [409] "fsl_fcs_meat"                                              
#> [410] "fsl_fcs_veg"                                               
#> [411] "fsl_fcs_fruit"                                             
#> [412] "fsl_fcs_oil"                                               
#> [413] "fsl_fcs_sugar"                                             
#> [414] "fsl_fcs_condiments"                                        
#> [415] "fsl_rcsi_note"                                             
#> [416] "fsl_rcsi_lessquality"                                      
#> [417] "fsl_rcsi_borrow"                                           
#> [418] "fsl_rcsi_mealsize"                                         
#> [419] "fsl_rcsi_mealadult"                                        
#> [420] "fsl_rcsi_mealnb"                                           
#> [421] "fsl_hhs_nofoodhh"                                          
#> [422] "fsl_hhs_nofoodhh_freq"                                     
#> [423] "fsl_hhs_sleephungry"                                       
#> [424] "fsl_hhs_sleephungry_freq"                                  
#> [425] "fsl_hhs_alldaynight"                                       
#> [426] "fsl_hhs_alldaynight_freq"                                  
#> [427] "fsl_source_food"                                           
#> [428] "other_fsl_source_food"                                     
#> [429] "fsl_source_food_ranked"                                    
#> [430] "fsl_lcsi_note"                                             
#> [431] "fsl_lcsi_stress1"                                          
#> [432] "fsl_lcsi_stress2"                                          
#> [433] "fsl_lcsi_stress3"                                          
#> [434] "fsl_lcsi_stress4"                                          
#> [435] "fsl_lcsi_crisis1"                                          
#> [436] "fsl_lcsi_crisis2"                                          
#> [437] "fsl_lcsi_crisis3"                                          
#> [438] "fsl_lcsi_emergency1"                                       
#> [439] "fsl_lcsi_emergency2"                                       
#> [440] "fsl_lcsi_emergency3"                                       
#> [441] "fsl_lcsi_other_reason"                                     
#> [442] "other_fsl_lcsi_other_reason"                               
#> [443] "fsl_lcsi_en_stress1"                                       
#> [444] "fsl_lcsi_en_stress2"                                       
#> [445] "fsl_lcsi_en_stress3"                                       
#> [446] "fsl_lcsi_en_stress4"                                       
#> [447] "fsl_lcsi_en_crisis1"                                       
#> [448] "fsl_lcsi_en_crisis2"                                       
#> [449] "fsl_lcsi_en_crisis3"                                       
#> [450] "fsl_lcsi_en_emergency1"                                    
#> [451] "fsl_lcsi_en_emergency2"                                    
#> [452] "fsl_lcsi_en_emergency3"                                    
#> [453] "fsl_lcsi_en_other_reason"                                  
#> [454] "other_fsl_lcsi_en_other_reason"                            
#> [455] "fsl_liv_act"                                               
#> [456] "fsl_crop_type"                                             
#> [457] "other_fsl_crop_type"                                       
#> [458] "fsl_crop_area_change"                                      
#> [459] "fsl_crop_diff"                                             
#> [460] "other_fsl_crop_diff"                                       
#> [461] "fsl_crop_harv_change"                                      
#> [462] "note_crop_harv_change"                                     
#> [463] "fsl_livestock_type"                                        
#> [464] "other_fsl_livestock_type"                                  
#> [465] "grp_sub_livestock"                                         
#> [466] "grp_oxen"                                                  
#> [467] "fsl_oxen_n_now"                                            
#> [468] "fsl_oxen_n_ly"                                             
#> [469] "grp_oxen"                                                  
#> [470] "fsl_oxen_n_change"                                         
#> [471] "other_fsl_oxen_n_change"                                   
#> [472] "fsl_oxen_diff"                                             
#> [473] "other_fsl_oxen_diff"                                       
#> [474] "grp_camel"                                                 
#> [475] "fsl_camel_n_now"                                           
#> [476] "fsl_camel_n_ly"                                            
#> [477] "grp_camel"                                                 
#> [478] "fsl_camel_n_change"                                        
#> [479] "other_fsl_camel_n_change"                                  
#> [480] "fsl_camel_diff"                                            
#> [481] "other_fsl_camel_diff"                                      
#> [482] "grp_cattle"                                                
#> [483] "fsl_cattle_n_now"                                          
#> [484] "fsl_cattle_n_ly"                                           
#> [485] "grp_cattle"                                                
#> [486] "fsl_cattle_n_change"                                       
#> [487] "other_fsl_cattle_n_change"                                 
#> [488] "fsl_cattle_diff"                                           
#> [489] "other_fsl_cattle_diff"                                     
#> [490] "grp_horse"                                                 
#> [491] "fsl_horse_n_now"                                           
#> [492] "fsl_horse_n_ly"                                            
#> [493] "grp_horse"                                                 
#> [494] "fsl_horse_n_change"                                        
#> [495] "other_fsl_horse_n_change"                                  
#> [496] "fsl_horse_diff"                                            
#> [497] "other_fsl_horse_diff"                                      
#> [498] "grp_mule"                                                  
#> [499] "fsl_mule_n_now"                                            
#> [500] "fsl_mule_n_ly"                                             
#> [501] "grp_mule"                                                  
#> [502] "fsl_mule_n_change"                                         
#> [503] "other_fsl_mule_n_change"                                   
#> [504] "fsl_mule_diff"                                             
#> [505] "other_fsl_mule_diff"                                       
#> [506] "grp_donkey"                                                
#> [507] "fsl_donkey_n_now"                                          
#> [508] "fsl_donkey_n_ly"                                           
#> [509] "grp_donkey"                                                
#> [510] "fsl_donkey_n_change"                                       
#> [511] "other_fsl_donkey_n_change"                                 
#> [512] "fsl_donkey_diff"                                           
#> [513] "other_fsl_donkey_diff"                                     
#> [514] "grp_sheep"                                                 
#> [515] "fsl_sheep_n_now"                                           
#> [516] "fsl_sheep_n_ly"                                            
#> [517] "grp_sheep"                                                 
#> [518] "fsl_sheep_n_change"                                        
#> [519] "other_fsl_sheep_n_change"                                  
#> [520] "fsl_sheep_diff"                                            
#> [521] "other_fsl_sheep_diff"                                      
#> [522] "grp_goat"                                                  
#> [523] "fsl_goat_n_now"                                            
#> [524] "fsl_goat_n_ly"                                             
#> [525] "grp_goat"                                                  
#> [526] "fsl_goat_n_change"                                         
#> [527] "other_fsl_goat_n_change"                                   
#> [528] "fsl_goat_diff"                                             
#> [529] "other_fsl_goat_diff"                                       
#> [530] "grp_poultry"                                               
#> [531] "fsl_poultry_n_now"                                         
#> [532] "fsl_poultry_n_ly"                                          
#> [533] "grp_poultry"                                               
#> [534] "fsl_poultry_n_change"                                      
#> [535] "other_fsl_poultry_n_change"                                
#> [536] "fsl_poultry_diff"                                          
#> [537] "other_fsl_poultry_diff"                                    
#> [538] "grp_sub_livestock"                                         
#> [539] "grp_food_storing"                                          
#> [540] "fsl_food_storing"                                          
#> [541] "other_fsl_food_storing"                                    
#> [542] "grp_food_storing"                                          
#> [543] "fsl"                                                       
#> [544] "cm"                                                        
#> [545] "cm_note"                                                   
#> [546] "cm_income_total"                                           
#> [547] "cm_income_sources_reduced_1"                               
#> [548] "other_cm_income_sources_reduced_1"                         
#> [549] "cm_income_sources_reduced_2"                               
#> [550] "other_cm_income_sources_reduced_2"                         
#> [551] "cm_expenditure_frequent"                                   
#> [552] "other_cm_expenditure_frequent"                             
#> [553] "cm_expenditure_frequent_note"                              
#> [554] "cm_expenditure_frequent_food"                              
#> [555] "cm_expenditure_frequent_rent"                              
#> [556] "cm_expenditure_frequent_water"                             
#> [557] "cm_expenditure_frequent_nfi"                               
#> [558] "cm_expenditure_frequent_utilities"                         
#> [559] "cm_expenditure_frequent_fuel"                              
#> [560] "cm_expenditure_frequent_transportation"                    
#> [561] "cm_expenditure_frequent_communication"                     
#> [562] "cm_expenditure_frequent_other"                             
#> [563] "cm_expenditure_infrequent"                                 
#> [564] "other_cm_expenditure_infrequent"                           
#> [565] "cm_expenditure_infrequent_note"                            
#> [566] "cm_expenditure_infrequent_shelter"                         
#> [567] "cm_expenditure_infrequent_clothing"                        
#> [568] "cm_expenditure_infrequent_nfi"                             
#> [569] "cm_expenditure_infrequent_health"                          
#> [570] "cm_expenditure_infrequent_education"                       
#> [571] "cm_expenditure_infrequent_debt"                            
#> [572] "cm_expenditure_infrequent_other"                           
#> [573] "cm_market_time"                                            
#> [574] "cm_market_barriers_access"                                 
#> [575] "other_cm_market_barriers_access"                           
#> [576] "cm_market_barriers_purchase"                               
#> [577] "other_cm_market_barriers_purchase"                         
#> [578] "cm"                                                        
#> [579] "health"                                                    
#> [580] "health_note"                                               
#> [581] "health_ind"                                                
#> [582] "health_ind_pos"                                            
#> [583] "health_person_id"                                          
#> [584] "health_ind_age"                                            
#> [585] "health_ind_gender"                                         
#> [586] "health_ind_name"                                           
#> [587] "health_ind_gender_fr"                                      
#> [588] "health_ind_healthcare_needed"                              
#> [589] "health_ind_healthcare_needed_y"                            
#> [590] "health_ind_healthcare_received"                            
#> [591] "health_ind_healthcare_needed_type"                         
#> [592] "other_health_ind_healthcare_needed_type"                   
#> [593] "health_ind_healthcare_received_location"                   
#> [594] "other_health_ind_healthcare_received_location"             
#> [595] "health_ind"                                                
#> [596] "health_ind_healthcare_needed_n"                            
#> [597] "health_facility_time"                                      
#> [598] "health_barriers"                                           
#> [599] "other_health_barriers"                                     
#> [600] "health_barriers_main"                                      
#> [601] "other_health_barriers_main"                                
#> [602] "health"                                                    
#> [603] "nutrition"                                                 
#> [604] "nut_ind_note"                                              
#> [605] "nut_ind"                                                   
#> [606] "nut_ind_pos"                                               
#> [607] "nut_person_id"                                             
#> [608] "nut_ind_age"                                               
#> [609] "nut_ind_gender"                                            
#> [610] "nut_ind_name"                                              
#> [611] "nut_ind_age_0_4"                                           
#> [612] "nut_ind_under5_age_months"                                 
#> [613] "nut_ind_age_vitamina"                                      
#> [614] "nut_ind_age_deworming"                                     
#> [615] "nut_ind_under5_sick_yn"                                    
#> [616] "nut_ind_under5_sick_symptoms"                              
#> [617] "other_nut_ind_under5_sick_symptoms"                        
#> [618] "nut_ind_under5_sick_location"                              
#> [619] "other_nut_ind_under5_sick_location"                        
#> [620] "nut_ind_under5_vitamina"                                   
#> [621] "nut_ind_under5_deworming"                                  
#> [622] "nut_ind"                                                   
#> [623] "nut_treat_acute_malnutrition"                              
#> [624] "nutrition"                                                 
#> [625] "protection"                                                
#> [626] "prot_needs_intro"                                          
#> [627] "prot_needs_1_services"                                     
#> [628] "prot_needs_1_services_threats"                             
#> [629] "prot_needs_1_services_population"                          
#> [630] "prot_needs_1_justice"                                      
#> [631] "prot_needs_1_justice_threats"                              
#> [632] "prot_needs_1_justice_population"                           
#> [633] "prot_needs_2_activities"                                   
#> [634] "prot_needs_2_activities_threats"                           
#> [635] "prot_needs_2_activities_population"                        
#> [636] "prot_needs_2_social"                                       
#> [637] "prot_needs_2_social_threats"                               
#> [638] "prot_needs_2_social_population"                            
#> [639] "prot_needs_3_movement"                                     
#> [640] "other_prot_needs_3_movement"                               
#> [641] "prot_needs_3_movement_threats"                             
#> [642] "prot_needs_3_movement_population"                          
#> [643] "prot_concern_introduction"                                 
#> [644] "prot_concern_hh_note"                                      
#> [645] "prot_concern_freq_cope"                                    
#> [646] "prot_concern_freq_hlp"                                     
#> [647] "prot_concern_freq_displaced"                               
#> [648] "prot_concern_freq_home_crime"                              
#> [649] "prot_concern_hh_freq_violence"                             
#> [650] "prot_concern_hh_freq_explosive"                            
#> [651] "prot_concern_hh_freq_kidnapping"                           
#> [652] "prot_concern_hh_freq_forcedrecruitment"                    
#> [653] "prot_concern_hh_freq_discrimination"                       
#> [654] "prot_concern_freq_gbv_walking"                             
#> [655] "prot_concern_freq_gbv_areas"                               
#> [656] "prot_concern_freq_gbv_areas_type"                          
#> [657] "prot_concern_impact"                                       
#> [658] "prot_child_sep"                                            
#> [659] "prot_child_sep_girls_n"                                    
#> [660] "prot_child_sep_boys_n"                                     
#> [661] "prot_child_sep_reason"                                     
#> [662] "prot_child_sep_reason_other"                               
#> [663] "prot_ind"                                                  
#> [664] "prot_ind_pos"                                              
#> [665] "prot_person_id"                                            
#> [666] "prot_ind_age"                                              
#> [667] "prot_ind_gender"                                           
#> [668] "prot_ind_name"                                             
#> [669] "prot_child_work"                                           
#> [670] "prot_child_labour"                                         
#> [671] "prot_ind"                                                  
#> [672] "ch_pr_behaviour_change"                                    
#> [673] "other_ch_pr_behaviour_change"                              
#> [674] "prot_id_yn"                                                
#> [675] "prot_id_who"                                               
#> [676] "prot_id_missing_reason"                                    
#> [677] "prot_id_missing_reason_other"                              
#> [678] "prot_legal_yn"                                             
#> [679] "prot_legal_yn_other"                                       
#> [680] "prot_legal_barriers"                                       
#> [681] "prot_legal_barriers_other"                                 
#> [682] "prot_legal_barriers_who"                                   
#> [683] "prot_adult_sep"                                            
#> [684] "prot_adult_sep_reason"                                     
#> [685] "prot_adult_sep_reason_other"                               
#> [686] "prot_adult_sep_mechanism"                                  
#> [687] "prot_mines"                                                
#> [688] "prot_mines_other"                                          
#> [689] "prot_services_availability"                                
#> [690] "durable_solutions"                                         
#> [691] "ds_plans"                                                  
#> [692] "ds_plans_timeline"                                         
#> [693] "durable_solutions"                                         
#> [694] "shocks_risks"                                              
#> [695] "hazard_env_type"                                           
#> [696] "hazard_sp_type"                                            
#> [697] "shocks_risks"                                              
#> [698] "gender_inclusion"                                          
#> [699] "gi_responsible_note2"                                      
#> [700] "gi_responsible_farming"                                    
#> [701] "gi_responsible_caring"                                     
#> [702] "gi_responsible_water"                                      
#> [703] "gi_responsible_firewood"                                   
#> [704] "gi_responsible_chores"                                     
#> [705] "gi_responsible_assistance"                                 
#> [706] "gi_responsible_food"                                       
#> [707] "gi_responsible_nonfood"                                    
#> [708] "gi_responsible_clothes"                                    
#> [709] "gender_inclusion"                                          
#> [710] "consented"                                                 
#> [711] "location"                                                  
#> [712] "household_geopoint"                                        
#> [713] "location"                                                  
#> [714] "final_note"                                                
#> [715] "final_text"                                                

# Count questions
length(xlsform_questions(form))
#> [1] 715